moved to cleaner raw ast impl
